+++ resourcet-extra/CHANGELOG.md
@@ -1,5 +1,7 @@
 # Version 0.1.1
 
+* Add `acquire1`, `acquireType1`.
+
 * Safer `mkAcquire1` and `mkAcquireType1` internals.